+++ src/taskEvents/index.html
@@ -15,8 +15,6 @@
     />
     <link rel="stylesheet" href="style.css" />
     <script src="script.js" type="module" defer></script>
-    <script src="utils.js" type="module" defer></script>
-    <script src="helper.js" type="module" defer></script>
   </head>
   <body>
     <div class="overlay">
