+++ src/taskEvents/utils.js
@@ -1,5 +1,46 @@
 const BASE_URL = 'http://localhost:4000';
 
+function createElement({ type, attributes = {}, innerText }) {
+  const element = document.createElement(type);
+  Object.keys(attributes).forEach((item) => {
+    element.setAttribute(item, attributes[item]);
+  });
+  element.textContent = innerText;
+  return element;
+}
+
+function addLoader(container) {
+  const loader = createElement({
+    type: 'div',
+    attributes: { class: 'loader' },
+  });
+  const loaderP = createElement({
+    type: 'p',
+    attributes: { class: 'loaderP' },
+    innerText: 'Loading...',
+  });
+  loader.appendChild(loaderP);
+  container.appendChild(loader);
+}
+
+function removeLoader() {
+  document.querySelector('.loader').remove();
+}
+
+function addErrorMessage(container) {
+  const errorDiv = createElement({
+    type: 'div',
+    attributes: { class: 'error-div' },
+  });
+  const errorMsg = createElement({
+    type: 'p',
+    attributes: { class: 'error-message' },
+    innerText: 'Something went wrong!',
+  });
+  errorDiv.appendChild(errorMsg);
+  container.appendChild(errorDiv);
+}
+
 async function getTaskLogs(username) {
   let url;
   if (username) {
@@ -57,4 +98,27 @@ async function getSelfDetails() {
   return self_details;
 }
 
-export { getTaskLogs, getTaskData, getUserData, getSelfDetails };
+async function getData(data) {
+  const message = data.body.message;
+  const userName = data.meta.username;
+  const taskId = data.meta.taskId;
+  const { taskData } = await getTaskData(taskId);
+
+  return {
+    ...taskData,
+    message,
+    userName,
+  };
+}
+
+export {
+  createElement,
+  addLoader,
+  addErrorMessage,
+  getTaskLogs,
+  getTaskData,
+  getUserData,
+  getSelfDetails,
+  getData,
+  removeLoader,
+};
